Split GPU and CPU test targets with dedicated runner scripts
Separate GPU-dependent tests from CPU-only tests so each category
can be run independently. GPU tests live under tests/gpu/ directories
and are invoked via `--test gpu`; CPU tests keep their original names.

- Move cutile-compiler inline GPU tests to tests/gpu/cuda_tile_runtime.rs
- Split cutile error_quality.rs into cpu (tests/error_quality.rs) and
  gpu (tests/gpu/error_quality.rs) variants
- Add run_cpu_tests.sh, run_gpu_tests.sh, and shared test_runner_common.sh
- Update run_all_tests.sh to call both scripts
