+++ cutile-compiler/tests/gpu.rs
@@ -1,0 +1,9 @@
+/*
+ * SPDX-FileCopyrightText: Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+//! Entry point for GPU-dependent cutile-compiler integration tests.
+
+#[path = "gpu/cuda_tile_runtime.rs"]
+mod cuda_tile_runtime;
